+++ Whatnow-Api/src/main/kotlin/com/depromeet/whatnow/WhatnowApiApplication.kt
@@ -2,7 +2,6 @@ package com.depromeet.whatnow
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories
 
 @SpringBootApplication
